Сделано главное меню, добавлены сотрудники
