+++ WebStore062020/Startup.cs
@@ -18,7 +18,7 @@ namespace WebStore062020
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddControllersWithViews();
+            services.AddControllersWithViews().AddRazorRuntimeCompilation();
         }
