+++ WebStore062020/Views/Employees/Details.cshtml
@@ -1,0 +1,31 @@
+﻿@model WebStore062020.Models.Employee
+
+    <div class="container">
+
+        <h3>Сотрудник</h3>
+
+        <dl class="dl-horizontal">
+            <dt>Id</dt>
+            <dd>@Model.Id</dd>
+
+            <dt>Фамилия</dt>
+            <dd>@Model.SurName</dd>
+
+            <dt>Имя</dt>
+            <dd>@Model.FirstName</dd>
+
+            <dt>Отчество</dt>
+            <dd>@Model.SecondName</dd>
+
+            <dt>Возраст</dt>
+            <dd>@Model.Age</dd>
+
+            <dt>Дата рождения</dt>
+            <dd>@Model.Birthday</dd>
+
+        </dl>
+
+
+        <hr />
+        <a asp-controller="Employees" asp-action="Index">К списку сотрудников</a>
+    </div>
