+++ WebStore062020/Views/Employees/Index.cshtml
@@ -1,0 +1,34 @@
+﻿@model IEnumerable<WebStore062020.Models.Employee>
+
+
+    <div class="container">
+        <hr />
+
+        <table class="table">
+            <thead>
+                <tr>
+                    <th>Фамилия</th>
+                    <th>Имя</th>
+                    <th>Отчество</th>
+                    <th>Возраст</th>
+                    <th></th>
+                </tr>
+            </thead>
+            <tbody>
+                @foreach (var employee in Model)
+                {
+                    <tr>
+                        <td>@employee.SurName</td>
+                        <td>@employee.FirstName</td>
+                        <td>@employee.SecondName</td>
+                        <td>@employee.Age</td>
+                        <td><a asp-controller="Employees" asp-action="Details" asp-route-id="@employee.Id">Информация</a></td>
+                    </tr>
+                }
+            </tbody>
+        </table>
+
+        <hr />
+
+        <a href="~/">На главную</a>
+    </div>
