--- WebStore062020/Views/Home/EmployeeInfo.cshtml
@@ -1,40 +1,0 @@
-﻿@model WebStore062020.Models.Employee
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Информация о сотруднике: @Model.SurName @Model.FirstName.Substring(0, 1) + ". " + @Model.SecondName.Substring(0, 1) + "."</title>
-</head>
-<body>
-
-    <h3>Сотрудник</h3>
-
-    <dl>
-        <dt>Id</dt>
-        <dd>@Model.Id</dd>
-
-        <dt>Фамилия</dt>
-        <dd>@Model.SurName</dd>
-
-        <dt>Имя</dt>
-        <dd>@Model.FirstName</dd>
-
-        <dt>Отчество</dt>
-        <dd>@Model.SecondName</dd>
-
-        <dt>Возраст</dt>
-        <dd>@Model.Age</dd>
-
-        <dt>Дата рождения</dt>
-        <dd>@Model.Birthday</dd>
-
-    </dl>
-
-
-    <hr />
-    <a href="~/Home/Employees">К списку сотрудников</a>
-
-</body>
-</html>
