--- WebStore062020/Views/Home/Employees.cshtml
@@ -1,45 +1,0 @@
-﻿@model IEnumerable<WebStore062020.Models.Employee>
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Сотрудники</title>
-</head>
-<body>
-
-    <h3>Сотрудники</h3>
-
-    <hr />
-
-    <table>
-        <thead>
-            <tr>
-                <th>Фамилия</th>
-                <th>Имя</th>
-                <th>Отчество</th>
-                <th>Возраст</th>
-                <th></th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var employee in Model)
-            {
-                <tr>
-                    <td>@employee.SurName</td>
-                    <td>@employee.FirstName</td>
-                    <td>@employee.SecondName</td>
-                    <td>@employee.Age</td>
-                    <td><a href="~/Home/EmployeeInfo/@employee.Id">Информация</a></td>
-                </tr>
-            }
-        </tbody>
-    </table>
-
-    <hr />
-
-    <a href="~/">На главную</a>
-
-</body>
-</html>
